+++ Bot.py
@@ -1,23 +1,26 @@
 import discord
 from discord.ext import commands
-from discord.ext.commands import Cog, Command
 import sys
 import dotenv
 import os
 import platform
 import time
 import argparse
-from typing import Dict, List, Optional
 
 import logging
 from rich.console import Console
 
 from Utils.Globals import PREFIX
-from Cogs.General import General
-from Utils.Helper import Emojis, INI, get_required_permissions
+
+from Utils.Helper import Emojis, INI
 from Utils.Utils import *
 
-COMMANDS_PATH = os.path.join(os.getcwd(), "Cogs")
+COMMAND_NAMES = [ # i change from auto get to this because i want to change the order of the categories
+    "General",
+    "Moderation",
+    "Utilities",
+    #... i will add more here later ;)
+]
 LOG_FILENAME = "Silly.log"
 
 parser = argparse.ArgumentParser()
@@ -41,196 +44,18 @@ args = parser.parse_args()
 
 
 class Silly(commands.Bot):
+    __version__ = "0.0.6"
     async def load_all_extensions(self) -> None:
         """Load all extensions from the Cogs folder.
         """
-        self.Logger.info(f"Loading commands from {COMMANDS_PATH}")
-        for filename in os.listdir(COMMANDS_PATH):
-            if filename.endswith(".py") and filename != "__init__.py":
-                try:
-                    await self.load_extension(f"Cogs.{filename[:-3]}")
-                    self.Logger.info(f"LOADED command {filename}")
-                except Exception as e:
-                    self.Logger.error(f"FAILED to load command {filename}: {e}")
+        self.Logger.info(f"Loading commands from \"Cogs\" folder")
+        for filename in COMMAND_NAMES:
+            try:
+                await self.load_extension(f"Cogs.{filename}")
+                self.Logger.info(f"LOADED command {filename}")
+            except Exception as e:
+                self.Logger.exception(f"FAILED to load command {filename}")
 
-    class Help(commands.HelpCommand):
-        def __init__(self, **options):
-            self.Emojis = Emojis
-            super().__init__(**options)
-            
-            
-            # i dont know why but help command doesnt support customize help coommand
-            # so i have to do this
-            self.command_attrs["description"] = "Need help, *hooman*? I can **smeowll** it!!\nI will **explain everything** I understand for **you**!"
-            self.command_attrs["brief"] = "Let the **silliest cat** help you"
-            self.command_attrs["aliases"] = ["meow, wheresisthesilliestcat", "pleasehelpimtoosilly"]
-            
-        
-        """def get_command_signature(self, command):
-            return f"{self.Helpers.Prefix}{command.qualified_name} {command.signature}"""
-            
-        # ━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━ >help ━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━ #
-        async def send_bot_help(self, mapping: Dict[Optional[Cog], List[Command]]):
-            ctx = self.context
-            embed = discord.Embed(
-                description=f"""# {self.Emojis.random()} ・ Meo Help Center
-**Meow!, I am _Silly_.**
-I am here to help you with your *silly* needs.
-""",
-                color=discord.Color.green()
-            )
-            embed.set_image(url="https://cdn.discordapp.com/attachments/1124562179635556362/1362386665569779803/Silly_6.gif?ex=680234f4&is=6800e374&hm=c82b1ace8d364a1b641a17f1b7b08b548e566af3f38e3c8f537464d44e82db3d&")
-            embed.set_thumbnail(url="https://cdn.discordapp.com/attachments/1124562179635556362/1362385462278225971/SillyCat.png?ex=680233d6&is=6800e256&hm=59610ec0e8157398191d3063463ff8deae66618787058c0e47318982c9a3775e&")
-            embed.add_field(name="Current Prefix", value=f"`{PREFIX}`", inline=False)
-            for cog, commands_list in mapping.items():
-                filtered = await self.filter_commands(commands_list, sort=True) # Filter commands and sort that shit
-                if filtered:
-                    Category = f"{cog.Emoji}   {cog.qualified_name}" if cog else f"{self.Emojis.UIA_Spinning}   Other (No Category)"
-                    embed.description += f"\n# {Category}\n"
-                    for cmdidx, cmd in enumerate(filtered):
-                        
-                        CommandName = cmd.name
-                        CommandDescription = cmd.brief
-                        
-                        # bro its chatgpt icl but i wrote comments
-                        (RequiredParams, OptionalParams) = ([f"[{param.name}]" for param in cmd.clean_params.values() # Get clean params (filtered the self and ctx)
-                                          if param.default is param.empty # Check if param has a default value, if not, it's required
-                                          and param.kind in (param.POSITIONAL_OR_KEYWORD, param.KEYWORD_ONLY) # Check if param is positional or keyword only
-                                          ],
-                                                            
-                                          [f"<{param.name}>" for param in cmd.clean_params.values() # Get clean params (filtered the self and ctx)
-                                          if param.default is not param.empty # Check if param has a default value, if yes, it's mot required
-                                          and param.kind in (param.POSITIONAL_OR_KEYWORD, param.KEYWORD_ONLY) # Check if param is positional or keyword only
-                                          ] # Get optional params
-                        )
-                        
-                        embed.description += f"""### {cmdidx+1}. {CommandName}{f" `{' '.join(RequiredParams)}`" if RequiredParams else ''}{f" `{' '.join(OptionalParams)}`" if OptionalParams else ''}{f"\n-#  Aliases: **{'**, **'.join(cmd.aliases)}**" if cmd.aliases else ''}
-> {repr(CommandDescription).strip('"').strip("'")}   
-""" 
-            await ctx.send(embed=embed)
-            
-    
-        # ------------------------------ >help <command> ----------------------------- #
-        async def send_command_help(self, command: commands.Command):
-            ctx = self.context
-            
-            # Get required permissions
-            required_permissions = get_required_permissions(command)
-            
-            permissions_text = (
-                "\n".join([f"- `{perm.replace('_', ' ').title()}`" for perm in required_permissions])
-                if required_permissions
-                else f"{self.Emojis.UIA_Spinning * 3}"
-            )
-            
-            embed = discord.Embed(
-                description=f"""# {self.Emojis.random()} ・ Meo Help Center
-**Meow!, I am _Silly_.**
-I am here to help you with your **_silly_ needs**.
-
-**Help for `{command.name}` in {command.cog.Emoji} __{command.cog_name or "Other"}__**
-
-# {self.Emojis.NerdCat}   Description
-{command.description}
-
-# {self.Emojis.JustWokeUp}   Usage
-{command.usage}
-# {self.Emojis.SecurityCat}   Required Permissions
-{permissions_text}
-
-# {self.Emojis.InsaneArhh}   Aliases
-{', '.join(command.aliases) if command.aliases else f'{self.Emojis.UIA_Spinning}'}
-""",
-                color=discord.Color.green()
-            )
-            embed.set_thumbnail(url="https://c.tenor.com/KO80NCIjQAUAAAAd/tenor.gif")
-            embed.set_image(url="https://cdn.discordapp.com/attachments/1124562179635556362/1362386665569779803/Silly_6.gif?ex=680234f4&is=6800e374&hm=c82b1ace8d364a1b641a17f1b7b08b548e566af3f38e3c8f537464d44e82db3d&")
-            await ctx.send(embed=embed)
-        
-        # ------------------------------ >help <category> ----------------------------- #
-
-        async def send_cog_help(self, group: commands.Cog):
-            ctx = self.context
-            embed = discord.Embed(
-                description=f"""# {self.Emojis.random()} ・ Meo Help Center
-**Meow!, I am _Silly_.**
-I am here to help you with your **_silly_ needs**.
-
-# {group.Emoji} __{group.qualified_name}__
-""",
-                color=discord.Color.green()
-            )
-            
-            #logging.getLogger("silly").info(f"Sending group help for {group.qualified_name}")
-            filtered = await self.filter_commands(group.get_commands(), sort=True)
-            for cmd in filtered:
-                CommandName = cmd.name
-                CommandDescription = cmd.brief
-                
-
-                (RequiredParams, OptionalParams) = ([f"[{param.name}]" for param in cmd.clean_params.values() # Get clean params (filtered the self and ctx)
-                                    if param.default is param.empty # Check if param has a default value, if not, it's required
-                                    and param.kind in (param.POSITIONAL_OR_KEYWORD, param.KEYWORD_ONLY) # Check if param is positional or keyword only
-                                    ],
-                                                    
-                                    [f"<{param.name}>" for param in cmd.clean_params.values() # Get clean params (filtered the self and ctx)
-                                    if param.default is not param.empty # Check if param has a default value, if yes, it's mot required
-                                    and param.kind in (param.POSITIONAL_OR_KEYWORD, param.KEYWORD_ONLY) # Check if param is positional or keyword only
-                                    ] # Get optional params
-                )
-                
-                
-                embed.description += f"""- **{CommandName}{f" `{' '.join(RequiredParams)}`" if RequiredParams else ''}{f" `{' '.join(OptionalParams)}`" if OptionalParams else ''}**{f"\n-#  Aliases: **{'**, **'.join(cmd.aliases)}**" if cmd.aliases else ''}
-  - {repr(CommandDescription).strip('"').strip("'")}
-  
-"""
-            embed.set_thumbnail(url="https://c.tenor.com/KO80NCIjQAUAAAAd/tenor.gif")
-            embed.set_image(url="https://cdn.discordapp.com/attachments/1124562179635556362/1362386665569779803/Silly_6.gif?ex=680234f4&is=6800e374&hm=c82b1ace8d364a1b641a17f1b7b08b548e566af3f38e3c8f537464d44e82db3d&")
-            await ctx.send(embed=embed)
-            
-            
-        # ━━━━━━━━━━━━━━━━━━━━━━━━━━━━ ERROR WHEN COMMAND NOT FOUND ━━━━━━━━━━━━━━━━━━━━━━━━━━━━ #
-        async def send_error_message(self, error): 
-            ctx = self.context
-            
-            embed = discord.Embed(
-                description=f"""# {self.Emojis.random()} ・ Meo Help Center
-**Meow!, I am _Silly_.**
-I am here to help you with your **_silly_ needs**.
-### {self.Emojis.CatScare} ・ Error
-- {error}
-
-**Please use `{PREFIX}help` to get help.**
--# *im so proud*
-""",    
-                color=discord.Colour.red()
-            )
-            embed.set_image(url="https://cdn.discordapp.com/attachments/1124562179635556362/1362386665569779803/Silly_6.gif?ex=680234f4&is=6800e374&hm=c82b1ace8d364a1b641a17f1b7b08b548e566af3f38e3c8f537464d44e82db3d&")
-            embed.set_thumbnail(url="https://cdn.discordapp.com/attachments/1218440992651218959/1255879183931019294/image_2024-06-27_213452161.gif?ex=6801ac62&is=68005ae2&hm=e4bd704f6e03141cb1ff88f306e285654f41de4c45f666f0d79969317e17ce41&")
-            await ctx.send(embed=embed, delete_after=15)
-            
-        # ---------------------------------------------------------------------------- #
-        #                       REPLACED COMMAND CALLBACK                              #
-        # ---------------------------------------------------------------------------- #
-        async def command_callback(self, ctx, *, command=None):
-            """The actual implementation of the help command."""
-            if command is None:
-                return await self.send_bot_help(self.get_bot_mapping())
-            
-            # Check if it's a cog name (case-insensitive)
-            command_lower = command.lower()
-            for cog in ctx.bot.cogs.values():
-                if cog.qualified_name.lower() == command_lower:
-                    return await self.send_cog_help(cog)
-            
-            # Check if it's a command (case-insensitive)
-            cmd = ctx.bot.get_command(command_lower)
-            if cmd is not None:
-                return await self.send_command_help(cmd)
-            
-            # If we get here, the command wasn't found
-            return await self.send_error_message(commands.CommandNotFound(f"**Uh oh, the command named `{command}` was not found!**"))
-            
     
     def __init__(self) -> None:
         # Init console n Logger
@@ -266,7 +91,7 @@ I am here to help you with your **_silly_ needs**.
   
   
 Made with [bright_red]<3[/] by [royal_blue1]fat0426[/] on [cornflower_blue]Discord[/]
-[gray35]https://github.com/pojo1807
+[gray35]https://github.com/pojo1807 [Current version: {self.__version__}]
 Python     v{platform.python_version()}
 Discord.py v{discord.__version__}[/]""", highlight=False)
         
@@ -274,12 +99,13 @@ Discord.py v{discord.__version__}[/]""", highlight=False)
             
         self.Emojis = Emojis
         
-        
         super().__init__(intents=discord.Intents.all(), # i dont know but the "all" works instead of "default"
-                         #help_command=None, # remove the default help command
+                         help_command=None,
                          command_prefix=PREFIX
                          )
-        
+
+    
+
     # ━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━ #
     #                                             ON READY                                             #
     # ━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━ #
@@ -292,8 +118,6 @@ f"Logged in as \"{self.user.name}#{self.user.discriminator}\" [gray35]({self.use
         )
         # Set status
         await self.change_presence(activity=discord.Game(name="with my hooman")) # static for now
-        
-        self.help_command = self.Help()
         
         if self.Emojis._emojis is None:
             self.Logger.warning(f"No emojis found! It may cause response issues.\nPlease add every single emoji in \"_Emojis\" folder in https://discord.com/developers/applications/{self.user.id}/emojis")
@@ -314,6 +138,15 @@ f"Logged in as \"{self.user.name}#{self.user.discriminator}\" [gray35]({self.use
             if args.sync == "all":
                 try:
                     sync_commands = await self.tree.sync()
+                    
+                    # test slash command style
+                    for cmd in sync_commands:
+                        #self.Logger.debug(f"cmd id: {cmd.id}")
+                        # save the id to the INI file
+                        # neu lan sau khong sync thi co the lay id tu file INI
+                        INI.set(f"Slash_Commands", cmd.name, str(cmd.id))
+                    with open("Settings.ini", "w") as f:
+                        INI.write(f)
                     
                     self.Logger.info(f"SYNCED {len(sync_commands)} commands!")
                 except Exception as e: